+++ CHANGELOG.md
@@ -52,6 +52,8 @@ All notable changes to this project will be documented in this file.
   - RFC-11: Onchain Activation ([#2302](https://github.com/malbeclabs/doublezero/pull/2302))
 - Monitor
   - Add link status to device-telemetry metrics to enable Grafana alerts to filter out links that are not in activated status
+- E2E tests
+  - Add GetLatency call to qaagent
 
 ## [v0.8.0](https://github.com/malbeclabs/doublezero/compare/client/v0.7.1...client/v0.8.0) – 2025-12-02